Fix bug where tile is inappropriately created
Previous approach naively mapped the area of interest to the entire tile.

This works if the user is zoomed in, but is not appropriate if the user zooms out to beyond the spatial area represented by the raster.

Also fixes a bug where the lons/lats are flipped by accident. Remember that Rasters.jl represents lons along the first dimension (i.e., the X dimension - the rows)