Update site identification and assessment process
Previously the polygon was rotated to align with the identified reef edge and only rotated a certain number of times to reduce computational cost.

However, the reef edge was based on the available outlines which are problematic in some cases as they may not represent the actual reef edge (when compared to the bathymetry data).

Instead, we rotate and assess the polygon in steps of 10 degrees to identify the approximately optimal rotation.